+++ Engine/Engine.vcxproj
@@ -72,10 +72,12 @@
     <ClInclude Include="Includes\camera\Controller.h" />
     <ClInclude Include="Includes\Engine.h" />
     <ClInclude Include="Includes\errors.h" />
+    <ClInclude Include="Includes\graphics\CameraUBO.h" />
     <ClInclude Include="Includes\graphics\Config.h" />
     <ClInclude Include="Includes\graphics\Geometry.h" />
     <ClInclude Include="Includes\graphics\Graphics.h" />
     <ClInclude Include="Includes\graphics\Math.h" />
+    <ClInclude Include="Includes\graphics\SceneUBO.h" />
     <ClInclude Include="Includes\graphics\Shaders.h" />
     <ClInclude Include="Includes\graphics\Vertex.h" />
     <ClInclude Include="Includes\Logger.h" />
