+++ Engine/Engine.vcxproj.filters
@@ -191,6 +191,12 @@
     <ClInclude Include="Includes\world\LightObject.h">
       <Filter>Engine\World</Filter>
     </ClInclude>
+    <ClInclude Include="Includes\graphics\CameraUBO.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="Includes\graphics\SceneUBO.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="..\assets\shaders\HLSL\no_texture_shader.frag">
