+++ Engine/Includes/camera/Controller.h
@@ -6,6 +6,8 @@
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/quaternion.hpp>
 
+#include <functional>
+
 #include "Camera.h"
 
 struct Actions
